common: Move blueprint struct to common
Share the same blueprint struct between the weldrapi and cloudapi info
handling. This is a subset of the blueprint and not suitable for
starting a compose or fully reproducing the original compose. Only use
it for the info output.

Related: RHEL-60147
